3.8

+++ SEGUNDO_T/spring_boot_web_mvc_jdbc_ventas-master/src/main/java/org/iesbelen/dao/ComercialDAOImpl.java
@@ -41,7 +41,7 @@ public class ComercialDAOImpl implements ComercialDAO {
 			ps.setString(idx++, comercial.getNombre());
 			ps.setString(idx++, comercial.getApellido1());
 			ps.setString(idx++, comercial.getApellido2());
-			ps.setFloat(idx++, comercial.getComision());
+			ps.setBigDecimal(idx++, comercial.getComision());
 			return ps;
 		},keyHolder);
 
@@ -58,7 +58,7 @@ public class ComercialDAOImpl implements ComercialDAO {
                 							  rs.getString("nombre"), 
                 							  rs.getString("apellido1"),
                 							  rs.getString("apellido2"), 
-                							  rs.getFloat("comisión"))
+                							  rs.getBigDecimal("comisión"))
                 						 	
         );
 		
@@ -76,7 +76,7 @@ public class ComercialDAOImpl implements ComercialDAO {
 						rs.getString("nombre"),
 						rs.getString("apellido1"),
 						rs.getString("apellido2"),
-						rs.getFloat("comisión")
+						rs.getBigDecimal("comisión")
 				),
 				id
 		);
